pyproject.toml
[project]
name = "recall-rag"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.10.0",
    "google-genai>=1.7.0",
    "pyyaml>=6.0.2",
]
